rebrand minimal example instead of creating a new
