+++ examples/minimal/client/index.ts
@@ -1,7 +1,7 @@
 import { createTRPCProxyClient, httpBatchLink } from '@trpc/client';
 import type { AppRouter } from '../server';
 
-const client = createTRPCProxyClient<AppRouter>({
+const trpc = createTRPCProxyClient<AppRouter>({
   links: [
     httpBatchLink({
       url: 'http://localhost:2022',
@@ -10,10 +10,17 @@ const client = createTRPCProxyClient<AppRouter>({
 });
 
 async function main() {
-  const result = await client.greet.query('tRPC');
+  /**
+   * Inferring types
+   */
+  const users = await trpc.userList.query();
+  //    ^?
 
-  // Type safe
-  console.log(result.greeting.toUpperCase());
+  const user = await trpc.userById.query('1');
+  //    ^?
+
+  const createdUser = await trpc.userCreate.mutate({ name: 'sachinraja' });
+  //    ^?
 }
 
-void main();
+main().catch(console.error);
