+++ examples/minimal/server/index.ts
@@ -1,25 +1,53 @@
 import { initTRPC } from '@trpc/server';
 import { createHTTPServer } from '@trpc/server/adapters/standalone';
-
-export type AppRouter = typeof appRouter;
+import { z } from 'zod';
 
 const t = initTRPC.create();
 
-const publicProcedure = t.procedure;
 const router = t.router;
+const publicProcedure = t.procedure;
+
+type User = { id: string; name: string };
+const db = {} as {
+  user: {
+    findMany: () => Promise<User[]>;
+    findById: (id: string) => Promise<User>;
+    create: (data: { name: string }) => Promise<User>;
+  };
+};
 
 const appRouter = router({
-  greet: publicProcedure
-    .input((val: unknown) => {
-      if (typeof val === 'string') return val;
-      throw new Error(`Invalid input: ${typeof val}`);
-    })
-    .query(({ input }) => ({ greeting: `hello, ${input}!` })),
+  userList: publicProcedure.query(async () => {
+    // Retrieve users from a datasource, this is an imaginary database
+    const users = await db.user.findMany();
+    //    ^?
+    return users;
+  }),
+  userById: publicProcedure.input(z.string()).query(async (opts) => {
+    const { input } = opts;
+    //      ^?
+    // Retrieve the user with the given ID
+    const user = await db.user.findById(input);
+    return user;
+  }),
+  userCreate: publicProcedure
+    .input(z.object({ name: z.string() }))
+    .mutation(async (opts) => {
+      const { input } = opts;
+      //      ^?
+      // Create a new user in the database
+      const user = await db.user.create(input);
+      //    ^?
+      return user;
+    }),
 });
 
-createHTTPServer({
+// Export type router type signature,
+// NOT the router itself.
+export type AppRouter = typeof appRouter;
+
+const server = createHTTPServer({
   router: appRouter,
-  createContext() {
-    return {};
-  },
-}).listen(2022);
+});
+
+server.listen(3000);
