--- examples/quickstart/README.md
@@ -1,19 +1,0 @@
-# A minimal working tRPC example
-
-Requires node 18 (for global fetch).
-
-## Playing around
-
-```
-npm i
-npm run dev
-```
-
-Try editing the ts files to see the type checking in action :)
-
-## Building
-
-```
-npm run build
-npm run start
-```
