--- examples/quickstart/client/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@examples/quickstart-client",
-  "version": "10.17.0",
-  "private": true,
-  "dependencies": {
-    "@trpc/client": "^10.17.0"
-  },
-  "devDependencies": {
-    "@types/node": "^18.7.20"
-  }
-}
