--- examples/quickstart/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@examples/quickstart",
-  "private": true,
-  "version": "10.17.0",
-  "scripts": {
-    "build": "tsc",
-    "dev:server": "tsx watch server",
-    "dev:client": "wait-port 2022 && tsx watch client",
-    "dev": "run-p dev:* --print-label",
-    "start": "node dist/server",
-    "lint": "eslint --ext \".js,.ts,.tsx\" --report-unused-disable-directives */*.ts",
-    "test-dev": "start-server-and-test 'tsx server' 2022 'tsx client'",
-    "test-start": "start-server-and-test 'node dist/server' 2022 'node dist/client'"
-  },
-  "workspaces": [
-    "client",
-    "server"
-  ],
-  "devDependencies": {
-    "@types/node": "^18.7.20",
-    "eslint": "^8.30.0",
-    "npm-run-all": "^4.1.5",
-    "start-server-and-test": "^1.12.0",
-    "tsx": "^3.12.3",
-    "typescript": "^4.8.3",
-    "wait-port": "^1.0.1"
-  }
-}
