--- examples/quickstart/server/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@examples/quickstart-server",
-  "version": "10.17.0",
-  "private": true,
-  "dependencies": {
-    "@trpc/server": "^10.17.0",
-    "zod": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^18.7.20"
-  }
-}
