--- examples/quickstart/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "esnext",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "strict": true,
-    "outDir": "dist"
-  },
-  "include": ["server", "client"]
-}
