+++ pnpm-workspace.yaml
@@ -5,6 +5,5 @@ packages:
   - 'examples/.test/*'
   - 'examples/minimal/*'
   - 'examples/minimal-react/*'
-  - 'examples/quickstart/*'
   - 'www'
   - 'www/og-image'
