+++ www/docs/main/quickstart.mdx
@@ -436,7 +436,7 @@ import clsx from 'clsx';
 >
   <Iframe
     src={
-      `https://stackblitz.com/github/trpc/trpc/tree/quickstart/examples/quickstart?` +
+      `https://stackblitz.com/github/trpc/trpc/tree/main/examples/minimal?` +
       searchParams({
         embed: '1',
         file: [
